+++ features/overlays/fishingProfitTracker.js
@@ -72,6 +72,16 @@ registerIf(
 );
 
 registerIf(
+    register("Chat", (mobNameText, event) => onShardsCharmed(mobNameText, 1)).setCriteria(triggers.CHARM_NAGA_SALT_SHARD_MESSAGE).setContains(),
+    () => settings.fishingProfitTrackerOverlay && isInSkyblock() && isInFishingWorld(getWorldName())
+);
+
+registerIf(
+    register("Chat", (mobNameText, shardsCount, event) => onShardsCharmed(mobNameText, +(shardsCount.removeFormatting()))).setCriteria(triggers.CHARM_NAGA_SALT_SHARDS_MESSAGE).setContains(),
+    () => settings.fishingProfitTrackerOverlay && isInSkyblock() && isInFishingWorld(getWorldName())
+);
+
+registerIf(
     register("Chat", (petDisplayName, level, event) => onPetReachedMaxLevel(+level, petDisplayName))
         .setCriteria(triggers.PET_LEVEL_UP_MESSAGE)
         .setContains(),
@@ -386,7 +396,7 @@ function onShardFished(shardText) {
         if (!isTrackerVisible() || !shardText || !isSessionActive) return;
 
         const [article, ...shardNameParts] = shardText.removeFormatting().split(' ');
-        const shardName = shardNameParts.join(' ');
+        const shardName = shardNameParts.join(' ') + ' Shard';
         refreshItemData(i => i.itemName === shardName.removeFormatting(), 1);
         refreshPrices();
         refreshOverlay();
@@ -406,7 +416,7 @@ function onShardCaughtInBlackHole(shardsText) {
 
         const [countText, ...shardNameParts] = shardsText.removeFormatting().split(' ');
         const count = countText === 'a' || countText === 'an' ? 1 : +(countText.replace('x', ''));
-        const shardName = shardNameParts.join(' ');
+        const shardName = shardNameParts.join(' ') + ' Shard';
 
         refreshItemData(i => i.itemName === shardName, count);
         refreshPrices();
@@ -414,6 +424,27 @@ function onShardCaughtInBlackHole(shardsText) {
     } catch (e) {
 		console.error(e);
 		console.log(`[FeeshNotifier] [ProfitTracker] Failed to track Shard caught in Black Hole.`);
+	}
+}
+
+/**
+ * 
+ * @param {string} mobNameText "a &fSea Archer", "an &aEnt"
+ * @param {number} shardsCount Charmed shards count to add to the tracker
+ */
+function onShardsCharmed(mobNameText, shardsCount) {
+    try {
+        if (!isTrackerVisible() || !mobNameText || !shardsCount || !isSessionActive) return;
+
+        const [article, ...mobNameParts] = mobNameText.removeFormatting().split(' ');
+        const shardName = mobNameParts.join(' ') + ' Shard';
+
+        refreshItemData(i => i.itemName === shardName, shardsCount);
+        refreshPrices();
+        refreshOverlay();
+    } catch (e) {
+		console.error(e);
+		console.log(`[FeeshNotifier] [ProfitTracker] Failed to track charmed Shard.`);
 	}
 }
 
@@ -540,7 +571,7 @@ function detectInventoryChanges() {
                 slotItemName += ` (${description})`;
             }
 
-            if (slotItemName === 'Fishing Exp Boost') {
+            if (slotItemName.endsWith('Exp Boost')) {
                 const loreLines = getLore(item);
                 const description = loreLines.find(line => line.endsWith('PET ITEM')).removeFormatting().split(' ')[0];
                 slotItemName += ` (${description})`;